web_backend/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
gensim==4.3.2
nltk==3.8.1
pyspellchecker==0.8.1
